requirements.txt
# Core ML/DL libraries
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0

# Audio Processing
librosa>=0.8.0
soundfile>=0.10.0

# NLP/Text Processing
transformers>=4.10.0
tokenizers>=0.10.0

# Data Processing
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
# pickle5>=0.0.11  # Not needed for Python 3.8+
# argparse  # Built into Python 3.2+
# pathlib  # Built into Python 3.4+

# Optional: For Kaggle dataset download
kaggle>=1.5.12

# Optional: For advanced audio features
opensmile>=2.0.0

# Optional: For video processing
ffmpeg-python>=0.2.0